// FootballManagement/football-ng/src/app/Model/match.ts
export class Match {
  id: number;
  happenedDate: Date;
  homeTeamId: number;
  homeTeamName: string;
  logoHome: string;
  visitTeamId: number;
  visitTeamName: string;
  logoVisit: string;
  tournamentId: number;
  homeGoals: number;
  visitGoals: number;
  locationId: number;
  locationName: string;
  locationImage: string;
  description: string;
  dayNo: string;
  finishPlayer: string;
  finishDetails: string;
}